FU-106 feat: get product reference images
