+++ src/services/customer/reservation.ts
@@ -49,3 +49,11 @@ export async function getFormBase(productId: string) {
   };
   return result;
 }
+
+export async function getImageList(photographerId: string) {
+  const response = await api
+    .get(`customer/product/images/${photographerId}`)
+    .json<{ data: string[] }>();
+  const { data } = response;
+  return data;
+}
